requirements.txt
pydantic>=2.0
pyyaml>=6.0
openai>=1.0
psycopg2-binary>=2.9
numpy>=1.24
requests>=2.28
